+++ Algebras/CRel/TypeConstructors/IntsTC.cpp
@@ -102,14 +102,14 @@ void IntsTC::Delete(const ListExpr typeExpr, Word &value)
   return DefaultDelete(typeExpr, value);
 }
 
-bool IntsTC::Open(SmiRecord &valueRecord, uint64_t &offset,
+bool IntsTC::Open(SmiRecord &valueRecord, size_t &offset,
                   const ListExpr typeExpr, Word &value)
 {
   return DefaultOpen(valueRecord, offset,
                                                       typeExpr, value);
 }
 
-bool IntsTC::Save(SmiRecord &valueRecord, uint64_t &offset,
+bool IntsTC::Save(SmiRecord &valueRecord, size_t &offset,
                   const ListExpr typeExpr, Word &value)
 {
   return DefaultSave(valueRecord, offset, typeExpr, value);
